Store NaN for non-hits in clusters. Modify code to adapt for that. Store clusters as float16

+++ lib/constants.py
@@ -25,7 +25,7 @@ dt_ci = numpy.dtype([
     ('TSPIDR', numpy.uint16),
     ('sumToT', numpy.uint16)
 ])
-dt_clusters = 'uint16'
+dt_clusters = 'float16'
 
 CLUSTER_CHUNK_SIZE = 100000
